Severstal/severstal-dashboard/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
faiss-cpu
openpyxl
